Added a TODOs to code.

+++ mulberry/font_creator.rb
@@ -3,7 +3,7 @@ require 'pathname'
 module Mulberry
   class FontCreator
 
-    # font creation works a little differently
+    # font creation works a little differently than other factories
     # instead of just creating a font file, they're supplying a path
     # to an existing font file, which we're going to copy into the assets folder
 
@@ -24,6 +24,10 @@ module Mulberry
       font_fname = File.basename(fontpath)
       font_name = font_fname.split('.')[0]
 
+      if File.exists? File.join(fonts_dir, font_fname)
+        raise "Font has already been added"
+      end
+
       # copy the font file over
       FileUtils.cp(fontpath, File.join(fonts_dir, font_fname))
       font_newpath = File.join(fonts_dir, font_fname)
@@ -35,10 +39,15 @@ module Mulberry
       FileUtils.mkdir_p themes_dir unless File.exists? themes_dir
 
       # write the font import into the file
+      # TODO: this should really be prepended to base.scss, as it won't work unless it's added
+      # ahead of the imports & etc...
       File.open(File.join(themes_dir, theme_cssfile), 'a') do |f|
         pathstring = Pathname.new("#{font_newpath}").relative_path_from(Pathname.new(themes_dir))
         f.write "@font-face {\n\tfont-family: '#{font_name}';\n\tsrc: url(#{pathstring}) format('truetype')\n\tfont-weight: normal;\n\tfont-style:normal;\n}\n"
       end
+
+      puts "Added #{font_name} to the app."
+      puts "Added @font-face import to the #{theme_cssfile}. You should customize it as necessary now."
 
     end
   end
